Move handlers pkg back to dashboard ; Move dashboard.go -> router.go
