+++ pkg/dashboard/dashboard.go
@@ -15,63 +15,159 @@
 package dashboard
 
 import (
-	"fmt"
+	"bytes"
+	"encoding/json"
+	"html/template"
 	"net/http"
 
 	packr "github.com/gobuffalo/packr/v2"
 	"github.com/gorilla/mux"
+	"k8s.io/klog"
 
-	"github.com/fairwindsops/goldilocks/pkg/dashboard/handlers"
+	"github.com/fairwindsops/goldilocks/pkg/dashboard/helpers"
 	"github.com/fairwindsops/goldilocks/pkg/summary"
 )
 
-var (
-	markdownBox = (*packr.Box)(nil)
+// templates
+const (
+	// MainTemplateName is the main template
+	MainTemplateName = "main.gohtml"
+	// HeadTemplateName contains styles and meta info
+	HeadTemplateName = "head.gohtml"
+	// NavbarTemplateName contains the navbar
+	NavbarTemplateName = "navbar.gohtml"
+	// PreambleTemplateName contains an empty preamble that can be overridden
+	PreambleTemplateName = "preamble.gohtml"
+	// DashboardTemplateName contains the content of the dashboard
+	DashboardTemplateName = "dashboard.gohtml"
+	// NamespaceTemplateName contains the content for a namespace
+	NamespaceTemplateName = "namespace.gohtml"
+	// ContainerTemplateName contains the content for a container
+	ContainerTemplateName = "container.gohtml"
+	// FooterTemplateName contains the footer
+	FooterTemplateName = "footer.gohtml"
+	// CheckDetailsTemplateName is a page for rendering details about a given check
+	CheckDetailsTemplateName = "check-details.gohtml"
 )
 
-// GetMarkdownBox returns a binary-friendly set of markdown files with error details
-func GetMarkdownBox() *packr.Box {
-	if markdownBox == (*packr.Box)(nil) {
-		markdownBox = packr.New("Markdown", "../../docs")
-	}
-	return markdownBox
+var templateBox = (*packr.Box)(nil)
+
+// templateData is passed to the dashboard HTML template
+type templateData struct {
+	// BasePath is the base URL that goldilocks is being served on, used in templates for html base
+	BasePath string
+
+	// VPAData is the summary information to display on the rendered dashboard
+	VPAData summary.Summary
+
+	// JSON is the json version of the summary data provided to the html window object for debugging and user interaction
+	JSON template.JS
 }
 
-// GetRouter returns a mux router serving all routes necessary for the dashboard
-func GetRouter(setters ...Option) *mux.Router {
-	opts := defaultOptions()
-	for _, setter := range setters {
-		setter(opts)
+// getTemplateBox returns a binary-friendly set of templates for rendering the dash
+func getTemplateBox() *packr.Box {
+	if templateBox == (*packr.Box)(nil) {
+		templateBox = packr.New("Templates", "templates")
 	}
+	return templateBox
+}
 
-	router := mux.NewRouter()
+// DashboardAllNamespaces replies with the rendered dashboard (on the basePath) for the summarizer
+func Dashboard(opts Options) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		vars := mux.Vars(r)
 
-	// health
-	router.Handle("/health", handlers.Health("OK"))
-	router.Handle("/healthz", handlers.Healthz())
+		var namespace string
+		if val, ok := vars["namespace"]; ok {
+			namespace = val
+		}
 
-	// assets
-	router.Handle("/favicon.ico", handlers.Asset("images/favicon-32x32.png"))
-	router.PathPrefix("/static/").Handler(handlers.StaticAssets("/static/"))
+		// TODO [hkatz] add caching or refresh button support
+		summarizer := summary.NewSummarizer(
+			summary.ForNamespace(namespace),
+			summary.ForVPAsWithLabels(opts.vpaLabels),
+			summary.ExcludeContainers(opts.excludedContainers),
+		)
 
-	// dashboard
-	summarizerAllNamespaces := summary.NewSummarizer(
-		summary.ForVPAsWithLabels(opts.vpaLabels),
-		summary.ExcludeContainers(opts.excludedContainers),
-	)
-	router.Handle("/dashboard", handlers.Dashboard(opts.basePath, summarizerAllNamespaces))
-
-	// root
-	router.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		// catch all other paths that weren't matched
-		if r.URL.Path != "/" {
-			http.NotFound(w, r)
+		vpaData, err := summarizer.GetSummary()
+		if err != nil {
+			klog.Errorf("Error getting vpaData: %v", err)
+			http.Error(w, "Error running summary.", http.StatusInternalServerError)
+			return
+		}
+		jsonData, err := json.Marshal(vpaData)
+		if err != nil {
+			http.Error(w, "Error serializing summary vpaData", http.StatusInternalServerError)
 			return
 		}
 
-		// default redirect on root path
-		http.Redirect(w, r, fmt.Sprintf("%s://%s:%d/dashboard", r.URL.Scheme, r.URL.Host, opts.port), http.StatusMovedPermanently)
+		data := templateData{
+			BasePath: opts.basePath,
+			VPAData:  vpaData,
+			JSON:     template.JS(jsonData),
+		}
+		tmpl, err := getTemplate("main")
+		if err != nil {
+			klog.Errorf("Error getting template data %v", err)
+			http.Error(w, "Error getting template data", http.StatusInternalServerError)
+			return
+		}
+
+		writeTemplate(tmpl, &data, w)
+	})
+}
+
+// getTemplate puts together the dashboard template. Individual pieces can be overridden before rendering.
+func getTemplate(name string) (*template.Template, error) {
+	tmpl := template.New(name).Funcs(template.FuncMap{
+		"printResource":  helpers.PrintResource,
+		"getStatus":      helpers.GetStatus,
+		"getStatusRange": helpers.GetStatusRange,
+		"resourceName":   helpers.ResourceName,
+		"getUUID":        helpers.GetUUID,
 	})
 
-	return router
+	templateFileNames := []string{
+		DashboardTemplateName,
+		NamespaceTemplateName,
+		ContainerTemplateName,
+		HeadTemplateName,
+		NavbarTemplateName,
+		PreambleTemplateName,
+		FooterTemplateName,
+		MainTemplateName,
+	}
+
+	return parseTemplateFiles(tmpl, templateFileNames)
+}
+
+func parseTemplateFiles(tmpl *template.Template, templateFileNames []string) (*template.Template, error) {
+	templateBox := getTemplateBox()
+	for _, fname := range templateFileNames {
+		templateFile, err := templateBox.Find(fname)
+		if err != nil {
+			return nil, err
+		}
+
+		tmpl, err = tmpl.Parse(string(templateFile))
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	return tmpl, nil
+}
+
+func writeTemplate(tmpl *template.Template, data *templateData, w http.ResponseWriter) {
+	buf := &bytes.Buffer{}
+	err := tmpl.Execute(buf, data)
+	if err != nil {
+		klog.Errorf("Error executing template: %v", err)
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+	_, err = buf.WriteTo(w)
+	if err != nil {
+		klog.Errorf("Error writing template: %v", err)
+	}
 }
